+++ packages/commands/push.js
@@ -120,6 +120,11 @@ async function putTermFile(config) {
   formData.append("updating", "terms_translations");
   formData.append("file", fs.createReadStream(config.file));
   formData.append("overwrite", "1");
+  
+  if (config.syncTerms && config.sourceLang && config.language === config.sourceLang) {
+    formData.append('sync_terms', '1');
+  }
+
   if (config.pushParams) {
     Object.keys(config.pushParams)
       .forEach(paramName => formData.set(paramName, config.pushParams[paramName]))